add handleRequest
