+++ .npmignore
@@ -13,3 +13,4 @@ node_modules
 bash/benchmarks
 bin
 tests
+*.sh
